Serve Vite build with Express
