+++ package.json
@@ -1,21 +1,15 @@
 {
-  "name": "digi-backend",
+  "name": "digi-app",
   "version": "1.0.0",
+  "private": true,
   "type": "module",
-  "main": "server.js",
+  "engines": { "node": ">=18" },
   "scripts": {
-    "build": "npm run build --prefix frontend",
-    "start": "node server.js"
+    "build:frontend": "cd frontend && npm ci && npm run build && cp -R dist ../dist",
+    "build": "npm run build:frontend",
+    "start": "node server.mjs"
   },
   "dependencies": {
-    "axios": "^1.8.4",
-    "base-64": "^1.0.0",
-    "connect-mongo": "^5.1.0",
-    "dotenv": "^16.5.0",
-    "express": "^4.18.2",
-    "express-session": "^1.18.1",
-    "mongodb": "^6.15.0",
-    "morgan": "^1.10.0",
-    "nodemailer": "^6.10.1"
+    "express": "^4.19.2"
   }
 }
