+++ render.yaml
@@ -1,0 +1,9 @@
+services:
+  - type: web
+    name: digi-backend
+    env: node
+    plan: free
+    buildCommand: npm run build
+    startCommand: npm start
+    autoDeploy: true
+    healthCheckPath: /healthz
